road_semantics: almost finish the topological part;

+++ smart-ros-pkg/experimental/road_semantics/src/topo_graph.hpp
@@ -21,7 +21,7 @@ public:
 	class edge {
 	public:
 		vector<CvPoint> points;
-		size_t head_nodeCluster, end_nodeCluster;
+		int head_nodeCluster, end_nodeCluster;
 	};
 
 	vector<CvPoint> nodes;
